Changes the hook entry point for LibVirt - fixes #5
The hook point was moved up by one step.
It's early enough for us to get SSH access and thus upload the binary in time.
Not that this auto installation is super relevant, since this plugin will be used with
landrush_ip_get 99% of the time, which installs when necessary too.